+++ OrderSystem/Models/OrderItem.cs
@@ -2,13 +2,13 @@ namespace OrderSystem.Models;
 
 public sealed class OrderItem
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
 
-    public int OrderId { get; set; }
+    public long OrderId { get; set; }
 
     public Order? Order { get; set; }
 
-    public int ProductId { get; set; }
+    public long ProductId { get; set; }
 
     public Product? Product { get; set; }
